BUG: fix use of is in DE callback
While using differential evolution I found that
the callback function did not work if the result
was not identical (in the `is` sense) to True.

This commit changes the condition to be identical
in the `==` sense. Critically, this allows the
user to work with numpy arrays in an intuitive
way, where previously they could not, because the
numpy True wasn't equivalent in the `is` sense
to the python True.

This commit also includes a test to demonstrate
the bug.